Design an alternative view (gb 1c3e0f47)
Create a view of our tasks that accomplishes the following:
1. Shows a stream of relationships between tasks (parent -> child -> child) etc.
2. As part of this it creates the "task" flow so that we can see merge points of tickets into another and branching out from there etc.
3. You can easily switch to and from this view
4. As tasks are "completed they move and collapse left or the view moves, left = past, right = future.
5. Make the design be organic and easy to follow complex streams of work.

Ask any clarifications on the design before you get started. For the design just output a "mockup" but you can build in the navigation to the "mockup" so we can approve the design that you will implement.  Attached an idea of what I am visualizing in terms of high level.  ![image.png](.vibe-images/871fed85-e81e-4418-8176-65caee884854.png)