Add project dropdown (gb 376324ff)
You should be able to quickly change projects in the header by selecting it from a drop down, saving clicks of going back and forth between the projects view etc. Come up with a design that fits our design system and apply it to the header and make it functional.